MAINT: move code from fn_ops to tensor_ops
